Docker support config via env vars (#49)
This commit introduces the option to use environment variables to
override yaml config file values. This makes it easier especially for
managing configuration in Docker. 

Passing on the final configuration values - merged from config.yaml
and env vars - to npm for building the front-end is a bit tricky. The
way it is done is, a special file (ui-config.yaml), is generated when
running `npm run build`, using the new cli command `./discuit
-inject-config`, which is then used by npm.