Optimize subscription feed fetch SQL query
The query to fetch the subscriptions feed of a user that was in use
before this commit was of the following form (with some parts not
relevant to this discussion removed):

  SELECT
    posts.id,
    posts.type,
    posts.public_id,
    -- etc.....
    post_votes.id IS NOT NULL,
    post_votes.up
  FROM posts
    LEFT OUTER JOIN post_votes ON
      post_votes.user_id = <viewer_id>
      AND post_votes.post_id = posts.id
  WHERE
    posts.deleted = FALSE
    AND community_id IN (
      (SELECT
        community_members.community_id
      FROM
        community_members
      WHERE
        community_members.user_id = <viewer_id>)
    )
  ORDER BY
    posts.id DESC
  LIMIT
    11

This query was slow, and it was significantly slow if the user had
subscribed to more than a few communities (taking seconds). However,
if you remove the LEFT OUTER JOIN from the query, it would return the
data instantaneously. Or, without doing that, if you remove the
`community_id IN ()` part of the WHERE clause it would return the
data instantaneously as well. But the LEFT OUTER JOIN combined with
the subquery in the WHERE clause turns this into an excruciatingly
slow query to execute in MariaDB.

I'm not sure how this query is executed in MariaDB, but the subquery
is not executed first - as we can see - and then its results taken as
a constant array to match against the parent SELECT, as the obvious
way of executing this query would be.

This commit fixes this by breaking up the query into two seperate
queries. One to fetch the communities the user is a member of. And
the other to fetch the posts of the subscriptions feed. Note that the
only difference this commit makes to the previous SQL query is
removing the subquery; the rest of the query is identical. The new
query is of the form:

  SELECT
    posts.id,
    posts.type,
    posts.public_id,
    -- etc.....
    post_votes.id IS NOT NULL,
    post_votes.up
  FROM posts
    LEFT OUTER JOIN post_votes ON
      post_votes.user_id = <viewer_id>
      AND post_votes.post_id = posts.id
  WHERE
    posts.deleted = FALSE
    AND community_id IN (
      <community_id_1>,
      <community_id_2>,
      <community_id_3>,
      -- etc
    )
  ORDER BY
    posts.id DESC
  LIMIT
    11