Fix service worker caching bad asset files
If a file the service is supposed to cache for some reason doesn't exist
on the server, then the response of the server, whatever that was - an
error or a 0 byte response - was still being cached, incorrectly. In
which case happened, the PWA would have been completely broken for that
client until it was updated again.

This commit fixes this error by not caching any response that doesn't
have a 2xx HTTP response or have a x-service-worker-cache HTTP header.