User welcome notification (#140)
This commit introduces a new notification type, welcome, which is
sent to new users sometime after they sign up. Clicking on this
notification sends users to a community, specified in the
welcomeCommunity field of the config.yaml.

Right now, this notification is sent only a minute after a user signs
up, but that value is used for testing purposes.