Announcement notifications (#142)
Introduces a new notification type to let all the users know of new
announcements on the site; essentially, by sending a notification for
the announcement post to all the users.

While working on this, however, it occurred to me how the
notification rendering currently works—that is, on the front-end,
using the structured objects returned from the server—is not ideal,
especially for a site that uses the Web Push API for native
notifications. The reason being that, in the current system, when a
new notification type is introduced, the service-worker.js needs to
be updated for the users to get the new notification via the Web Push
API; and that that updating doesn't happen—and there's no way to make
it happen—until the user visits the site. If the notification was
rendered in a standard form on the server, however, this issue
wouldn't arise, as in that case there wouldn't be notification types.