UI: Limit on the update-prompt display frequency
If the update-prompt (the "a new version of this app is available..."
prompt) has been shown to the user in the last 20 minutes, don't show
it again even if there's a new update.

This change is made because some iOS (Safari) users have been getting
the update prompt frequently without the app being updated on the
backend. This is a duck-tape solution while the real bug is being is
being tracked down.