New user account restrictions
This commit introduces two restrictions on newly created user accounts:
one, a points and an age requirement to restrict the ability of new
accounts to increment points of other accounts; and two, a points
requirement to post image and link posts.

The first restriction is added so that it's harder to get around the
second restriction.

Both of these restrictions can be configured in the config file.